Add pre-processor to filter error messages based on RFC; #server-1344
